+++ EX11Birdwatching/src/ee/taltech/iti0202/birdwatching/bird/BirdDataController.java
@@ -23,7 +23,7 @@ public class BirdDataController {
                 String[] splittedLine = line.split(",");
 
                 birds.add(new BirdBuilder().setSpecies(splittedLine[0]).setWeight(Double.parseDouble(splittedLine[1]))
-                        .setWingspan(Integer.parseInt(splittedLine[2]))
+                        .setWingspan(Double.parseDouble(splittedLine[2]))
                         .setSex(Bird.Sex.valueOf(splittedLine[3].toUpperCase()))
                         .setAge(Bird.Age.valueOf(splittedLine[4].toUpperCase())).createBird());
             }
